+++ lib/new_relic/init.ex
@@ -4,31 +4,6 @@ defmodule NewRelic.Init do
   def run() do
     verify_erlang_otp_version()
     init_collector_host()
-    setup_logs_in_context()
-  end
-
-  def setup_logs_in_context() do
-    # TODO: setup only if agent enabled & feature is on & Elixir 1.10+
-
-    Application.get_env(:new_relic_agent, :logs_in_context, :disabled)
-    |> case do
-      :disabled ->
-        :skip
-
-      :direct ->
-        :logger.add_primary_filter(
-          :nr_logs_in_context,
-          {&NewRelic.LogsInContext.primary_filter/2, %{mode: :direct}}
-        )
-
-      :forward ->
-        :logger.add_primary_filter(
-          :nr_logs_in_context,
-          {&NewRelic.LogsInContext.primary_filter/2, %{mode: :forward}}
-        )
-
-        Logger.configure_backend(:console, format: {NewRelic.LogsInContext, :format})
-    end
   end
 
   @erlang_version_requirement ">= 21.0.0"
